Drastically improve clarity of AA lines by using floating point coordinates.

+++ combustion_graphing/src/graph/function.rs
@@ -6,7 +6,7 @@ use common::num_utils::ScaleExt;
 ///
 /// [https://en.wikipedia.org/wiki/Linear_equation](https://en.wikipedia.org/wiki/Linear_equation)
 pub fn graph_linear_equation<F, L>(width: u32, height: u32, domain_x: Range<f64>, domain_y: Range<f64>, steps: usize, f: F, mut draw_line: L) where F: Fn(f64) -> f64,
-                                                                                                                                                    L: FnMut(i64, i64, i64, i64) {
+                                                                                                                                                    L: FnMut(f64, f64, f64, f64) {
     let dx = (domain_x.end - domain_x.start) / steps as f64;
 
     let mut x0 = domain_x.start;
@@ -22,11 +22,11 @@ pub fn graph_linear_equation<F, L>(width: u32, height: u32, domain_x: Range<f64>
         let on_graph = !(y0.is_nan() || y1.is_nan()) && (domain_y.contains(y0) || domain_y.contains(y1));
 
         if on_graph {
-            let px0 = x0.scale(domain_x.start, domain_x.end, 0.0, width as f64) as i64;
-            let py0 = y0.scale(domain_y.start, domain_y.end, 0.0, height as f64) as i64;
+            let px0 = x0.scale(domain_x.start, domain_x.end, 0.0, width as f64);
+            let py0 = y0.scale(domain_y.start, domain_y.end, 0.0, height as f64);
 
-            let px1 = x1.scale(domain_x.start, domain_x.end, 0.0, width as f64) as i64;
-            let py1 = y1.scale(domain_y.start, domain_y.end, 0.0, height as f64) as i64;
+            let px1 = x1.scale(domain_x.start, domain_x.end, 0.0, width as f64);
+            let py1 = y1.scale(domain_y.start, domain_y.end, 0.0, height as f64);
 
             draw_line(px0, py0, px1, py1);
         }
@@ -46,7 +46,7 @@ pub fn graph_linear_equation<F, L>(width: u32, height: u32, domain_x: Range<f64>
 ///
 /// [https://en.wikipedia.org/wiki/Parametric_equation](https://en.wikipedia.org/wiki/Parametric_equation)
 pub fn graph_parametric_equation<F, L>(width: u32, height: u32, domain_t: Range<f64>, domain_x: Range<f64>, domain_y: Range<f64>, steps: usize, f: F, mut draw_line: L) where F: Fn(f64) -> (f64, f64),
-                                                                                                                                                                              L: FnMut(i64, i64, i64, i64) {
+                                                                                                                                                                              L: FnMut(f64, f64, f64, f64) {
     let dt = (domain_t.end - domain_t.start) / steps as f64;
 
     let mut xy0 = f(domain_t.start);
@@ -66,11 +66,11 @@ pub fn graph_parametric_equation<F, L>(width: u32, height: u32, domain_t: Range<
                 (domain_x.contains(x1) && domain_y.contains(y1)));
 
         if on_graph {
-            let px0 = x0.scale(domain_x.start, domain_x.end, 0.0, width as f64) as i64;
-            let py0 = y0.scale(domain_y.start, domain_y.end, 0.0, height as f64) as i64;
+            let px0 = x0.scale(domain_x.start, domain_x.end, 0.0, width as f64);
+            let py0 = y0.scale(domain_y.start, domain_y.end, 0.0, height as f64);
 
-            let px1 = x1.scale(domain_x.start, domain_x.end, 0.0, width as f64) as i64;
-            let py1 = y1.scale(domain_y.start, domain_y.end, 0.0, height as f64) as i64;
+            let px1 = x1.scale(domain_x.start, domain_x.end, 0.0, width as f64);
+            let py1 = y1.scale(domain_y.start, domain_y.end, 0.0, height as f64);
 
             draw_line(px0, py0, px1, py1);
         }
@@ -89,7 +89,7 @@ pub fn graph_parametric_equation<F, L>(width: u32, height: u32, domain_t: Range<
 ///
 /// [https://en.wikipedia.org/wiki/Polar_coordinate_system](https://en.wikipedia.org/wiki/Polar_coordinate_system)
 pub fn graph_polar_equation<F, L>(width: u32, height: u32, x: f64, y: f64, rotation: f64, domain_x: Range<f64>, domain_y: Range<f64>, steps: usize, f: F, draw_line: L) where F: Fn(f64) -> f64,
-                                                                                                                                                                              L: FnMut(i64, i64, i64, i64) {
+                                                                                                                                                                              L: FnMut(f64, f64, f64, f64) {
     graph_parametric_equation(width, height, 0.0..1.0, domain_x, domain_y, steps, |t: f64| {
         let angle = t * 2.0 * ::std::f64::consts::PI;
